+++ src/types/surfaces/brand.ts
@@ -1,3 +1,5 @@
-export function brand<T extends 'Control'|'Introspection'|'Monitoring'|'Subscription'>(type: T): `@@Strongbus${T}Surface` {
+export function brand<
+  T extends 'Control' | 'Emitting' | 'Introspection' | 'Monitoring' | 'Subscription'
+>(type: T): `@@Strongbus${T}Surface` {
   return `@@Strongbus${type}Surface`;
 }